Relecture de code numéro 2 ajout de l'attribut de type AuctionHall dans la classe User.

+++ EasyBid/src/Env/Price.java
@@ -21,12 +21,8 @@ public class Price {
 		return current;
 	}
 	
-	
-	public void setValue(int value) {
+	public void setValue(double value) {
 		this.value=value;
-		
 	}
-	
-	
-	
+
 }
